+++ myems-api/reports/spacecost.py
@@ -503,8 +503,7 @@ class Reporting:
                                            " WHERE space_id = %s "
                                            "     AND energy_category_id = %s "
                                            "     AND start_datetime_utc >= %s "
-                                           "     AND start_datetime_utc < %s "
-                                           " ORDER BY start_datetime_utc ",
+                                           "     AND start_datetime_utc < %s ",
                                            (child_space['id'],
                                             energy_category_id,
                                             reporting_start_datetime_utc,