+++ CHANGELOG.md
@@ -7,11 +7,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - added subtotal to meter/offlinemeter/virtualmeter batch report 
 - added mqtt to data source protocol in myems-api
+- added tbl_template_files to myems_reporting_db
+- added tbl_integrators to myems_system_db
+- 
 ### Changed
 - 
 
 ### Fixed
--
+- removed invalid ORDER BY clause from myems-api
 
 ### Removed
 -